pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cyberthreats"
version = "0.1.0"
description = "MCP-powered cloud architecture security reviewer"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.49.0",
    "fastmcp>=3.1.0",
    "google-genai>=1.0.0",
    "langchain>=1.2.12",
    "langgraph>=1.1.2",
    "langsmith>=0.7.16",
    "mcp>=1.26.0",
    "openai>=2.26.0",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "streamlit>=1.55.0",
]

[tool.setuptools.packages.find]
where = ["src"]
